pyproject.toml
[build-system]
requires = ["uv-build"]
build-backend = "uv_build"

[project]
name = "python-scaffolding"
version = "0.1.0"
description = "Python Scaffolding"
requires-python = ">=3.13,<3.14"
dependencies = [
    "pydantic",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "black",
    "mypy",
    "pytest",
    "pytest-cov",
    "ruff",
]

[tool.ruff]
target-version = "py313"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--tb=short --import-mode=importlib"
